refactor(synapsys): split matcher.js, reduce complexity for quality gate
- Extract stage-3 content helpers into matcher-content.js
- Extract stage-4 exclude evaluators into matcher-excludes.js
- Lift _resolvePromptToken and _evaluatePreToolExcludes helpers to cap
  matchPrompt / matchPreTool cyclomatic complexity under 10
- Replace duplicated truthy and exclude-resolution logic in
  readMemoryFile with _truthy / _buildExcludeResolved helpers

No behavior change. Restores Quality (static code gate) to green by
addressing the only non-allowlisted violations introduced on this branch.
